+++ src/main/kotlin/com/emberjs/template/EmberLiveTemplatesProvider.kt
@@ -1,0 +1,12 @@
+package com.emberjs.template
+
+import com.intellij.codeInsight.template.impl.DefaultLiveTemplatesProvider
+
+class EmberLiveTemplatesProvider : DefaultLiveTemplatesProvider {
+    override fun getDefaultLiveTemplateFiles() = DEFAULT_TEMPLATES
+    override fun getHiddenLiveTemplateFiles() = arrayOf<String>()
+
+    companion object {
+        val DEFAULT_TEMPLATES = arrayOf("/com/emberjs/template/Ember.js");
+    }
+}